// src/util/type/melding-type.ts
export enum MeldingType {
	DIALOG_MELDING = 'DIALOG_MELDING',
	SYSTEM_MELDING = 'SYSTEM_MELDING'
}
export enum SystemMeldingType {
	UTKAST_OPPRETTET = 'UTKAST_OPPRETTET',
	BESLUTTER_PROSESS_STARTET = 'BESLUTTER_PROSESS_STARTET',
	BLITT_BESLUTTER = 'BLITT_BESLUTTER',
	BESLUTTER_HAR_GODKJENT = 'BESLUTTER_HAR_GODKJENT',
	TATT_OVER_SOM_BESLUTTER = 'TATT_OVER_SOM_BESLUTTER',
	TATT_OVER_SOM_VEILEDER = 'TATT_OVER_SOM_VEILEDER',
	SENDT_TIL_VEILEDER = 'SENDT_TIL_VEILEDER',
	SENDT_TIL_BESLUTTER = 'SENDT_TIL_BESLUTTER',
	BESLUTTER_PROSESS_AVBRUTT = 'BESLUTTER_PROSESS_AVBRUTT'
}